+++ src/backend/jitsi_magnify/core/utils.py
@@ -1,3 +1,0 @@
-"""
-Utils that can be useful throughout Jitsi magnify's core app
-"""