+++ src/backend/magnify/core/utils.py
@@ -1,0 +1,3 @@
+"""
+Utils that can be useful throughout Magnify's core app
+"""