+++ src/terraform/output.tf
@@ -1,3 +1,3 @@
 output "objectstorage_bucket_name" {
-  value = openstack_objectstorage_container_v1.jitsi_magnify_media_storage.name
+  value = openstack_objectstorage_container_v1.magnify_media_storage.name
 }